o/snapstate: move tests
