+++ overlord/snapstate/snapstate_update_test.go
@@ -45,6 +45,7 @@ import (
 	"github.com/snapcore/snapd/overlord"
 	"github.com/snapcore/snapd/overlord/assertstate"
 	"github.com/snapcore/snapd/overlord/auth"
+	userclient "github.com/snapcore/snapd/usersession/client"
 
 	// So it registers Configure.
 	_ "github.com/snapcore/snapd/overlord/configstate"
@@ -9076,4 +9077,83 @@ func (s *snapmgrTestSuite) TestConditionalAutoRefreshCreatesPreDownloadChange(c 
 	c.Assert(chgs[0].Status(), Equals, state.DoneStatus)
 
 	checkPreDownloadChange(c, chgs[1], "some-snap", snap.R(2))
+}
+
+func (s *snapmgrTestSuite) TestAutoRefreshCreatePreDownload(c *C) {
+	restore := snapstate.MockRefreshAppsCheck(func(si *snap.Info) error {
+		return snapstate.NewBusySnapError(si, []int{123}, nil, nil)
+	})
+	defer restore()
+
+	s.state.Lock()
+	defer s.state.Unlock()
+
+	si := &snap.SideInfo{RealName: "some-snap", SnapID: "some-snap-id", Revision: snap.R(1)}
+	snapst := &snapstate.SnapState{
+		Active:   true,
+		Sequence: []*snap.SideInfo{si},
+		Current:  si.Revision,
+		SnapType: string(snap.TypeApp),
+	}
+	snapstate.Set(s.state, "some-snap", snapst)
+	snapsup := &snapstate.SnapSetup{
+		SideInfo: &snap.SideInfo{RealName: "some-snap", SnapID: "some-snap-id", Revision: snap.R(2)},
+		Flags:    snapstate.Flags{IsAutoRefresh: true},
+	}
+
+	ts, err := snapstate.DoInstall(s.state, snapst, snapsup, 0, "", inUseCheck)
+
+	var busyErr *snapstate.TimedBusySnapError
+	c.Assert(errors.As(err, &busyErr), Equals, true)
+	refreshInfo := busyErr.PendingSnapRefreshInfo()
+	c.Check(refreshInfo, DeepEquals, &userclient.PendingSnapRefreshInfo{
+		InstanceName:  "some-snap",
+		TimeRemaining: snapstate.MaxInhibition,
+	})
+
+	tasks := ts.Tasks()
+	c.Assert(tasks, HasLen, 1)
+	c.Assert(tasks[0].Kind(), Equals, "pre-download-snap")
+	c.Assert(tasks[0].Summary(), testutil.Contains, "Pre-download snap \"some-snap\" (2) from channel")
+}
+
+func (s *snapmgrTestSuite) TestAutoRefreshBusySnapButOngoingPreDownload(c *C) {
+	restore := snapstate.MockRefreshAppsCheck(func(si *snap.Info) error {
+		return snapstate.NewBusySnapError(si, []int{123}, nil, nil)
+	})
+	defer restore()
+
+	s.state.Lock()
+	defer s.state.Unlock()
+
+	si := &snap.SideInfo{RealName: "some-snap", SnapID: "some-snap-id", Revision: snap.R(1)}
+	snapst := &snapstate.SnapState{
+		Active:   true,
+		Sequence: []*snap.SideInfo{si},
+		Current:  si.Revision,
+		SnapType: string(snap.TypeApp),
+	}
+	snapstate.Set(s.state, "some-snap", snapst)
+	snapsup := &snapstate.SnapSetup{
+		SideInfo: &snap.SideInfo{RealName: "some-snap", SnapID: "some-snap-id", Revision: snap.R(2)},
+		Flags:    snapstate.Flags{IsAutoRefresh: true},
+	}
+
+	// create ongoing pre-download
+	chg := s.state.NewChange("pre-download", "")
+	task := s.state.NewTask("pre-download-snap", "")
+	task.Set("snap-setup", snapsup)
+	chg.AddTask(task)
+	chg.SetStatus(state.DoStatus)
+
+	ts, err := snapstate.DoInstall(s.state, snapst, snapsup, 0, "", inUseCheck)
+
+	var busyErr *snapstate.TimedBusySnapError
+	c.Assert(errors.As(err, &busyErr), Equals, true)
+	refreshInfo := busyErr.PendingSnapRefreshInfo()
+	c.Check(refreshInfo, DeepEquals, &userclient.PendingSnapRefreshInfo{
+		InstanceName:  "some-snap",
+		TimeRemaining: snapstate.MaxInhibition,
+	})
+	c.Assert(ts, IsNil)
 }
